Trigger send webhook task on alert group change
