+++ engine/apps/webhooks/tasks/__init__.py
@@ -1,1 +1,2 @@
+from .alert_group_status_change import alert_group_created, alert_group_status_change
 from .send_webhook_event import execute_webhook, send_webhook_event  # noqa: F401
